pyproject.toml
[tool.poetry]
name = "grad"
version = "0.0.1"
description = "Creating Great Gradients"
authors = ["Marvin Straathof <marvinstraathof@gmail.com>"]

[tool.poetry.dependencies]
python = "^3.9"
Pillow = "^8.4.0"

[tool.poetry.dev-dependencies]
pytest = "^5.2"
black = "^21.11b1"
mypy = "^0.910"
pyproject-flake8 = "^0.0.1-alpha.2"
isort = "^5.10.1"

[tool.black]
line-length = 80

[tool.isort]
profile = "black"
line_length = 80

[tool.flake8]
max-line-length = 80

[build-system]
requires = ["poetry-core>=1.0.0"]
build-backend = "poetry.core.masonry.api"
